lambdas/get_prices_lambda/lambda_function.py
import json
import logging
import os
from datetime import datetime, timezone

import boto3
from botocore.exceptions import ClientError

# ロガーの設定
logger = logging.getLogger()
logger.setLevel(logging.INFO)

# DynamoDBクライアントの初期化
dynamodb = boto3.resource('dynamodb')
table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])

def lambda_handler(event, context):
    """
    Lambda関数のメインハンドラー
    """
    try:
        logger.info("Starting price scraping process")
        
        # 現在の日付を取得
        current_date = datetime.now(timezone.utc).date()
        
        # 当日のデータが存在するか確認
        if check_today_data(current_date):
            logger.info("Today's data already exists")
            return {
                'statusCode': 200,
                'body': json.dumps('Today\'s data already exists')
            }
        
        # スクレイピング処理を実行
        prices = scrape_prices()
        
        # データを保存
        save_prices(prices)
        
        return {
            'statusCode': 200,
            'body': json.dumps('Price scraping completed successfully')
        }
        
    except Exception as e:
        logger.error(f"Error in lambda_handler: {str(e)}")
        return {
            'statusCode': 500,
            'body': json.dumps(f'Error: {str(e)}')
        }

def check_today_data(current_date):
    """
    当日のデータが存在するか確認
    """
    try:
        response = table.query(
            KeyConditionExpression='model = :model',
            FilterExpression='begins_with(timestamp, :date)',
            ExpressionAttributeValues={
                ':model': 'iPhone 16 128GB',  # テスト用のモデル
                ':date': current_date.isoformat()
            }
        )
        return len(response['Items']) > 0
    except ClientError as e:
        logger.error(f"Error checking today's data: {str(e)}")
        raise

def scrape_prices():
    """
    価格情報をスクレイピング
    """
    # TODO: スクレイピングロジックの実装
    return []

def save_prices(prices):
    """
    価格情報をDynamoDBに保存
    """
    # TODO: 保存ロジックの実装
    pass
