requirements.txt
attrs==24.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
flask==3.1.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
playwright==1.51.0
pyee==12.1.1
PySocks==1.7.1
pytest-cov==6.1.1
pytest==7.4.3
python-dotenv==1.1.0
PyYAML==6.0.1
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio-websocket==0.12.2
trio==0.30.0
typing_extensions==4.13.2
urllib3>=1.26,<2.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
tenacity==8.2.3
pydantic==2.6.1
pytest-mock==3.12.0